Move IS_PLONE_5 constant to config module
