+++ src/collective/polls/config.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+from plone import api
+
+
 PROJECTNAME = 'collective.polls'
 PROFILE = PROJECTNAME + ':default'
 
@@ -7,3 +10,5 @@ MEMBERS_ANNO_KEY = 'voters_members_id'
 VOTE_ANNO_KEY = 'option.%02d'
 
 PERMISSION_VOTE = 'collective.polls: Vote'
+
+IS_PLONE_5 = api.env.plone_version().startswith('5')
